requirements.txt
pushbullet.py~=0.12.0
selenium~=4.11.2